01_Introduction and research
Project Title *

HardWorker: stay focused and achieve goals

Write your app’s name and tagline.

HardWoker

A village-life-game version of time focused App.
Write a short paragraph describing your app. Make sure to mention what the app does and
what problem it solves.

This App is a role-playing game. Users plan tasks on a to-do-list with a period of time. When the
user runs the timer in the App, the user starts to do the task, and the game character takes care
of the garden. After the timer, the user will see the garden has plants growing. The user uses
this App more, there will be more plants growing in the garden.

Describe four specific functions that your app will do. List these ideas in order of
importance, with the first idea being the most important.

1. User can set the timer to begin focusing on doing tasks;
2. Users can plan the tasks in the to-do list in a period of time;
3. Users will get rewarded in the game by continuously staying in time focusing in the real world;
4. Users can relax by exploring the game.

List 4 potential roles that are a part of your initial user target group. Then, expand on one
of the roles with a detailed profile description of one of these potential users.

1. Mr. and Mrs. Queen try to teach their game addicted child to stay in studying.
2. Jack wants to finish his painting before the due date but he loves to goofy around.
3. Lauren is 25 years old and tries to lose weight. She has a hard time being disciplined in
exercising.
4. AJ tries to quit his smoking addiction. He wants to hold off smoking longer and longer until he
quits smoking.

Come up with 4 client/designer goals for your app design. You don't have to assign these
to a category (political, social, etc). List your goals in order of importance, with the first
goal being the most important.

1. This App encourages users to stay focused on doing things.
2. This App helps users to achieve goals.
3. This App helps users to create better habits.
4. The client makes money when users pay for hidden options in the game.

Name and briefly describe 3 existing apps that are similar to your app idea.
The similar Apps are Study Bunny, Focus Plant, Forest: Stay focused.

1. Study Bunny - gain a better mood and more money for the rabbits in the game by adding
up duration time, so the rabbit has a smiling face and uses the money to buy decorations
for the house and other places.
2. Focus Plant - have plants to grow by adding up duration time. Users can choose seeds
and grow them on a field of land.
3. Forest: Stay focused - Duration of time triggers rain to water the user’s garden. The
longer the duration of time, the bigger drop in the rain to water the garden. The plants
grow faster.